refactor: move channels_v2/sender.py to channels/sender.py
