+++ apps/channels/channels_v2/api_channel.py
@@ -6,7 +6,6 @@ from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
-from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.channels_v2.stages.core import (
     BotInteractionStage,
     ChatMessageCreationStage,
@@ -22,6 +21,7 @@ from apps.channels.channels_v2.stages.core import (
 )
 from apps.channels.channels_v2.stages.terminal import ActivityTrackingStage, PersistenceStage
 from apps.channels.const import MESSAGE_TYPES
+from apps.channels.sender import ChannelSender
 from apps.chat.exceptions import ChannelException
 from apps.chat.models import Chat
 from apps.experiments.models import Experiment, Participant, SessionStatus
