+++ apps/channels/channels_v2/channel_base.py
@@ -38,9 +38,9 @@ from apps.teams.utils import current_team
 
 if TYPE_CHECKING:
     from apps.channels.channels_v2.callbacks import ChannelCallbacks
-    from apps.channels.channels_v2.sender import ChannelSender
     from apps.channels.datamodels import BaseMessage
     from apps.channels.models import ExperimentChannel
+    from apps.channels.sender import ChannelSender
     from apps.experiments.models import Experiment
 
 logger = logging.getLogger("ocs.channels")
