+++ apps/channels/channels_v2/connect_channel.py
@@ -6,9 +6,9 @@ from apps.channels.capabilities import ChannelCapabilities, PlatformConsentConfi
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
-from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.clients.connect_client import CommCareConnectClient
 from apps.channels.const import MESSAGE_TYPES
+from apps.channels.sender import ChannelSender
 from apps.chat.exceptions import ChannelException
 
 logger = logging.getLogger("ocs.channels")
