+++ apps/channels/channels_v2/email_channel.py
@@ -14,11 +14,11 @@ from django.db import IntegrityError
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.datamodels import _MAX_REFERENCES, RawAttachment, SkippedAttachment
 from apps.channels.datamodels import EmailMessage as EmailMessageDatamodel
 from apps.channels.models import ChannelPlatform, ExperimentChannel
+from apps.channels.sender import ChannelSender
 from apps.channels.utils import is_email_domain_allowed
 from apps.experiments.models import ExperimentSession
 from apps.files.content_type import detect_content_type
