+++ apps/channels/channels_v2/evaluation_channel.py
@@ -7,7 +7,6 @@ from apps.channels.channels_v2.api_channel import NoOpSender
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
-from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.channels_v2.stages.core import (
     ChatMessageCreationStage,
     EvalsBotInteractionStage,
@@ -17,6 +16,7 @@ from apps.channels.channels_v2.stages.core import (
 )
 from apps.channels.channels_v2.stages.terminal import ActivityTrackingStage, PersistenceStage
 from apps.channels.const import MESSAGE_TYPES
+from apps.channels.sender import ChannelSender
 from apps.chat.exceptions import ChannelException
 from apps.service_providers.tracing import TracingService
