+++ apps/channels/channels_v2/facebook_channel.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.models import ChannelPlatform
+from apps.channels.sender import ChannelSender
 
 if TYPE_CHECKING:
     from io import BytesIO
