+++ apps/channels/channels_v2/pipeline.py
@@ -13,10 +13,10 @@ from apps.service_providers.tracing import TraceInfo
 if TYPE_CHECKING:
     from apps.channels.capabilities import ChannelCapabilities
     from apps.channels.channels_v2.callbacks import ChannelCallbacks
-    from apps.channels.channels_v2.sender import ChannelSender
     from apps.channels.channels_v2.stages.base import ProcessingStage
     from apps.channels.datamodels import BaseMessage
     from apps.channels.models import ExperimentChannel
+    from apps.channels.sender import ChannelSender
     from apps.chat.bots import PipelineBot
     from apps.chat.models import ChatMessage
     from apps.experiments.models import Experiment, ExperimentSession, Participant, ParticipantData
