+++ apps/channels/channels_v2/whatsapp_channel.py
@@ -7,10 +7,10 @@ from typing import TYPE_CHECKING
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.channels_v2.stages.core import AttachmentHydrationStage
 from apps.channels.datamodels import WhatsAppMessage
 from apps.channels.models import ChannelPlatform
+from apps.channels.sender import ChannelSender
 from apps.chat.exceptions import ServiceWindowExpiredException
 from apps.files.models import File, FilePurpose
 from apps.service_providers.models import MessagingProviderType
